% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{cellcycle_reScoring}
\alias{cellcycle_reScoring}
\title{Re-calculated the cell cycle scoring of a seurat data set.}
\usage{
cellcycle_reScoring(
  object,
  SCTransform = T,
  vars_regress = c("percent.mito", "S.Score", "G2M.Score")
)
}
\arguments{
\item{object}{Seurat object}

\item{SCTransform}{Logical. If it is necessary to regress the cell cycle and other variables in vars_regress. Default TRUE.}

\item{vars_regress}{variables to regress. Includes cell cycle and mitochondrial DNA, but more can be added or it can be modified.}
}
\value{
Seurat object with calculated cell cycle scoring and regressed data.
}
\description{
Re-calculated the cell cycle scoring of a seurat data set.
}
