% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{clusterIDassign}
\alias{clusterIDassign}
\title{Automatic assigner of cluster name.}
\usage{
clusterIDassign(
  object,
  clusters.names,
  clusters.column = "SCT_snn_res.0.1",
  output = NULL,
  new_colName = "cell_type"
)
}
\arguments{
\item{object}{Seurat object}

\item{clusters.names}{list of names for each cluster in order}

\item{clusters.column}{metadata column containing the clusters to be named. Must have numbers. The names will be given in the order of the number/factors}

\item{output}{if an output with the cell information, including cluster name and number is desire please indicate output file}
}
\value{
Seurat object but with the cluster names assigned.
}
\description{
Automatic assigner of cluster name.
}
