% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{FeaturePlot_saver}
\alias{FeaturePlot_saver}
\title{Dimensional reduction plot for UMAP coloring the expression of marker genes given by the user. Creates PNG plots only for the indicated gene.}
\usage{
FeaturePlot_saver(
  object,
  genes = NULL,
  filePreffix,
  pt.size = NULL,
  order = TRUE,
  width = 5000,
  height = 3000
)
}
\arguments{
\item{object}{Seurat object}

\item{genes}{Genes to plot}

\item{pt.size}{Dot size}

\item{order}{Boolean determining whether to plot cells in order of expression. Can be useful if cells expressing given feature are getting buried.}

\item{width}{width}

\item{height}{height}
}
\value{
PNG file. UMAP plot showing the indicated. marker genes
}
\description{
Dimensional reduction plot for UMAP coloring the expression of marker genes given by the user. Creates PNG plots only for the indicated gene.
}
