% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{GenesDotplot}
\alias{GenesDotplot}
\title{Dotplot of marker genes. Creates PNG plots}
\usage{
GenesDotplot(
  object,
  filePreffix,
  genes,
  group.by = NULL,
  cols = c("lightgrey", "#cc0000"),
  width = 3000,
  height = 2000
)
}
\arguments{
\item{object}{Seurat object}

\item{filePreffix}{Preffix to use for the file name, including the path to the saving folder}

\item{group.by}{Name of the metadata column to group for.}

\item{cols}{Colors to scale the plot. Vector of two colors.}
}
\description{
Dotplot of marker genes. Creates PNG plots
}
