% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{markerGenesDotplot}
\alias{markerGenesDotplot}
\title{Dotplot of marker genes. Creates PNG plots}
\usage{
markerGenesDotplot(
  object,
  filePreffix,
  group.by = NULL,
  cols = c("lightgrey", "#cc0000")
)
}
\arguments{
\item{object}{Seurat object}

\item{filePreffix}{Preffix to use for the file name, including the path to the saving folder}

\item{group.by}{Name of the metadata column to group for.}

\item{cols}{Colors to scale the plot. Vector of two colors.}
}
\value{
PNG file. Dot plot of the marker genes of each cell type.
}
\description{
Dotplot of marker genes. Creates PNG plots
}
