% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{markerGenesPlot_MonoNK}
\alias{markerGenesPlot_MonoNK}
\title{Dimensional reduction plot for UMAP coloring the expression of pre-defined marker genes for the monocyte and NK subsets. Creates PNG plots}
\usage{
markerGenesPlot_MonoNK(
  object,
  filePreffix,
  pt.size = NULL,
  order = TRUE,
  cell_type = "Monocyte"
)
}
\arguments{
\item{object}{Seurat object}

\item{filePreffix}{Preffix to use for the file name, including the path to the saving folder}

\item{pt.size}{Dot size}

\item{order}{Boolean determining whether to plot cells in order of expression. Can be useful if cells expressing given feature are getting buried.}

\item{cell_type}{Monocyte or NK}
}
\value{
PNG files with the plots of the markers genes of Monocytes or NK cells. DCs included.
}
\description{
The gene panels are the following:
Monocyte: CD14, CD16 (FCGR3A), CD2, CD11b (ITGAM), CD31 (PECAM1), CD56 (NCAM1), CD62L (SELL), CD115 (CSF1R), CD192 (CCR2), CX3CR1, CXCR3, CXCR4,CD45 (PTPRC)
Classical: CD14++ CD16-, LYZ, CD45 (PTPRC), CD36(neg), CD64(FCGR1A, neg)
Non classical: CD14(dim) CD16++, MS4A7,CD11c (ITGAX)
Intermediate: CD14+, CD16+, LYZ, IFI30, CD74, CD86, HLA-DRA, S100A8, S100A10
}
\details{
NK cells: KLRB1, CD56 (NCAM1), CD62L (SELL), CD57 (B3GAT1), CD94/NKG2A (KLRC1), CD16 (FCGR3A), NKR-P1 (CD161, KLRB1), NKG2D (KLRK1, CD314), NCR, FCGR3A TYROBP CLIC3 NKY7 PRF1 TRBC2(neg), TRAC(neg)
Dendritic cells: absence of CD3 (T cell), CD14 (Monocyte), CD19 (B cell), CD56 (NCAM1, NK cells) and CD66b (CEACAM8, granulocytes)
cDC: BDCA-1 (CD1C), CD8, CD8alpha, CD11b (ITGAM), CD11c (ITGAX), CD103 (ITGAE), CD205 (LY75), MHCII (HLA-DRA)
pDC: BDCA-2 (CLEC4C), CD45RA (PTPRC), CD123 (IL3RA), ILT-7 (LILRA4),TLR7, TLR9, CD11clow (ITGAX), MHCII (HLA-DRA) low
}
