% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{metadataMetrics_gradientPlot}
\alias{metadataMetrics_gradientPlot}
\title{Plot the umaps coloring the desired metric as a gradient}
\usage{
metadataMetrics_gradientPlot(
  object,
  metrics = "nFeature_RNA",
  output_image,
  width = 22,
  height = 20,
  pt.size = 0.15,
  alpha = 0.7
)
}
\arguments{
\item{object}{Seurat object}

\item{metrics}{metric in metadata to plot with a color gradient}

\item{output_image}{file name of the image}

\item{width}{image width}

\item{height}{image height}
}
\value{
PNG file. UMAP plot.
}
\description{
Plot the umaps coloring the desired metric as a gradient
}
