% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{percent_mitoPlot}
\alias{percent_mitoPlot}
\title{Plot the umaps coloring the mitochondrial DNA \%.}
\usage{
percent_mitoPlot(object, output_image, width = 22, height = 20)
}
\arguments{
\item{object}{Seurat object}

\item{output_image}{file name of the image}

\item{width}{image width}

\item{height}{image height}
}
\value{
PNG file. UMAP plot.
}
\description{
Plot the umaps coloring the mitochondrial DNA \%.
}
