% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{pre_process}
\alias{pre_process}
\title{Pre_process a seurat object by applying different cut-offs defined by the user.}
\usage{
pre_process(data_dir, Project_name = "Seurat_object")
}
\arguments{
\item{data_dir}{directory with the matrix, barcodes and features. String.}

\item{nGenes_min}{minimun number of genes a cell must have in order to keep it. integer.}

\item{nCounts_min}{minimun number of counts a cell must have in order to keep it. integer.}

\item{PercentMitoDNA}{maximun percentage of mitoDNA a cell must have in order to keep it. integer. e.g. 10}
}
\value{
Filtered seurat object.
}
\description{
Pre_process a seurat object by applying different cut-offs defined by the user.
}
