% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{resolutionPlots}
\alias{resolutionPlots}
\title{Plot the resolution umaps. Resolutions required 0.1, 0.2, 0.3, 0.4, 0.5}
\usage{
resolutionPlots(object, reduction = "umap", output.image, integrated = F)
}
\arguments{
\item{object}{Seurat object}

\item{clusters.names}{list of names for each cluster in order}

\item{clusters.column}{metadata clumn containing the clusters to considere. Must have numbers.}

\item{output}{if an output with the cell information, including cluster name and number is desire please indicate output file}
}
\value{
PNG file. UMAP plots showing the clustering of the different resolutions.
}
\description{
Plot the resolution umaps. Resolutions required 0.1, 0.2, 0.3, 0.4, 0.5
}
